Perform parsing on command creation.
Previously, all Cmd subtypes required that the User call the
*Cmd::Parse() method to initialize the command object. This commit
updates each command object to include a Create() factory function that
returns either a valid command object or an error code. The benefit is
that if Create() returns a command object, it's guaranteed the User can
call the command's Execute() method safely.