Alias the return type of the Cmd Create() methods.
The Cmd Create() methods now return a std::expected where the first
template argument is a shared_ptr to the command type. Unfortunately,
make_shared() requires access to a non private/protected constructor.
The workarounds for this issue are ugly:
https://stackoverflow.com/questions/8147027/how-do-i-call-stdmake-shared-on-a-class-with-only-protected-or-private-const
To keep the code cleanish, the decision was made to make the
constructor's public. This allows a user to bypass Create() at their own
risk.